ObjSeq_ApplyLinkedObjectTransform: (s32)basePitch cast drops spurious extsh on s16 store (99.08->99.6)

+++ src/main/objseq.c
@@ -3265,7 +3265,7 @@ void ObjSeq_ApplyLinkedObjectTransform(u8* obj, u8* seqObj, u8* seq)
             if ((s8)((ObjSeqState*)seq)->movementState == 2)
             {
                 *(s16*)(seqObj + 0) =
-                    (s16)(basePitch + (s32)(
+                    (s16)((s32)basePitch + (s32)(
                         (f32)((ObjSeqState*)seq)->rotOffsetX * ((ObjSeqState*)seq)->posOffsetScale));
                 *(s16*)(seqObj + 2) =
                     (s16)(baseYaw + (s32)((f32)((ObjSeqState*)seq)->rotOffsetY * ((ObjSeqState*)seq)->posOffsetScale));
